requirements.txt
discord.py
asyncio
configparser
SQLAlchemy
dislash.py==1.4.1
wavelink
art
aiohttp
git+git://github.com/Rapptz/discord-ext-menus
humanize
requests
ujson
psycopg2
pycountry
six
google-cloud-translate>=3.0.0
tortoise-orm
google-cloud-vision
azure-cognitiveservices-vision-contentmoderator
msrest
asyncpg
pillow
requests_async
openai
